+++ CHANGELOG.rst
@@ -21,7 +21,6 @@ Added
 - server endpoint at ``DELETE /models`` to unload models from server memory
 - added count vector featurizer to create bag of words representation
 - added embedding intent classifier implemented in tensorflow
-- added tensorflow_embedding pipeline
 - added tensorflow requirements
 
 Changed
@@ -39,6 +38,8 @@ Changed
 - Ignore hidden files when listing projects or models
 - Docker Images now run on python 3.6 for better non-latin character set support
 - updated requirements_full.txt to include tensorflow requirements
+- updated documentation, added info about tensorflow_embedding pipeline
+- updated registry, added tensorflow_embedding pipeline
 
 Removed
 -------